+++ apps/dot-components-e2e/src/integration/AppToolbar/AppToolbar.spec.ts
@@ -1,7 +1,5 @@
 describe('dot-components: AppToolbar component', () => {
-  beforeEach(() =>
-    cy.visit('/iframe.html?id=experimental-apptoolbar--default')
-  );
+  beforeEach(() => cy.visit('/iframe.html?id=components-apptoolbar--default'));
 
   it('should have a dot- prefix', () => {
     cy.get('header').should('have.class', 'dot-app-toolbar');