+++ libs/dot-components/src/lib/components/app-toolbar/AppToolbar.stories.tsx
@@ -1,41 +1,130 @@
 import React from 'react';
+import { Input } from '@material-ui/core';
 import { DotAppToolbar, AppToolbarProps } from './AppToolbar';
+import { DotAvatar } from '../avatar/Avatar';
+import { DotSidebar } from '../sidebar/Sidebar';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
 export default {
-  title: 'Experimental/AppToolbar',
+  title: 'Components/AppToolbar',
   component: DotAppToolbar,
   argTypes: {
-    brand: {
-      defaultValue: 'Project Lisbon',
+    avatar: {
+      defaultValue: (
+        <DotAvatar alt="Batman" text="BW" size="small" type="text" />
+      ),
+    },
+    borderColor: {
+      defaultValue: '#1abc9c',
+    },
+    appName: {
+      defaultValue: 'Batman',
     },
     children: {
-      defaultValue: <input type="text" placeholder="search" />,
+      defaultValue: <Input type="text" placeholder="search" />,
     },
     items: {
       defaultValue: [
         {
           iconId: 'notification-bell',
           title: 'Alerts',
-          url: '/',
+          type: 'button',
         },
         {
           iconId: 'help',
           title: 'Help',
-          url: '/',
+          type: 'button',
         },
         {
           iconId: 'apps',
           title: 'Admin',
-          url: '/',
+          type: 'button',
         },
         {
           iconId: 'dark',
-          onClick: () => console.log('theme btn clicked'),
+          iconSize: 'small',
+          onClick: () => console.log('theme clicked'),
+          text: 'Theme',
           title: 'Toggle Theme',
-          url: '/',
+          type: 'button',
         },
       ],
+    },
+    mainMenu: {
+      defaultValue: (
+        <DotSidebar
+          navItems={[
+            {
+              iconId: 'process-template',
+              text: 'Progressions',
+              url: '/progressions',
+            },
+            {
+              iconId: 'satellite-group',
+              text: 'Pipelines',
+              url: '/pipelines',
+            },
+            {
+              iconId: 'dashboard',
+              text: 'Insights',
+              url: '/insights',
+            },
+            {
+              text: 'Workflow',
+              type: 'divider',
+            },
+            {
+              iconId: 'block',
+              text: 'Packages',
+              url: '/packages',
+            },
+            {
+              iconId: 'flag',
+              text: 'Features',
+              url: '/features',
+            },
+            {
+              iconId: 'collection',
+              text: 'Projects',
+              url: '/projects',
+            },
+            {
+              iconId: 'file-lines',
+              text: 'Workitems',
+              url: '/workitems',
+            },
+            {
+              iconId: 'change',
+              text: 'Changes',
+              url: '/changes',
+            },
+            {
+              iconId: 'square-settings',
+              text: 'Artifacts',
+              url: '/artifacts',
+            },
+            {
+              text: 'System',
+              type: 'divider',
+            },
+            {
+              iconId: 'monitor-gears',
+              text: 'Tasks',
+              url: '/tasks',
+            },
+            {
+              iconId: 'archive',
+              text: 'Buckets',
+              url: '/buckets',
+            },
+            {
+              iconId: 'monitor',
+              text: 'Environment',
+              url: '/environment',
+            },
+          ]}
+        />
+      ),
     },
   },
 } as Meta;